Завершение тестов по алгебраичским алгоритмам.

+++ AlgebraicAlgorithms/Fibonacci/FibonacciB.cs
@@ -7,9 +7,9 @@ namespace TestApp
         public string Title { get => "2b. Через итерацию"; }
         public string Run(string[] data)
         {
-            var n = Convert.ToInt32(data[0]);
-            long a = 0;
-            long b = 1;
+            var n = Convert.ToInt64(data[0]);
+            ulong a = 0;
+            ulong b = 1;
 
             for(int i = 0; i < n;i++) {
                 var c = b;
